data_pipeline/requirements.txt
pandas>=2.2.0
numpy>=1.26.0
pyarrow>=15.0.0
requests>=2.31.0
python-dotenv>=1.0.0
RateMyProfessorAPI>=1.3.6
rapidfuzz>=3.6.0
tqdm>=4.66.0
anthropic>=0.39.0
openai>=1.40.0
jsonschema>=4.22.0
supabase>=2.8.0
scikit-learn>=1.5.0
xgboost>=2.1.0
shap>=0.46.0
optuna>=3.6.0
pulp>=2.9.0
matplotlib>=3.9.0
seaborn>=0.13.0
ipykernel>=6.29.0
jupyter>=1.0.0
